+++ display_books.rb
@@ -1,0 +1,9 @@
+require_relative 'app'
+
+module Display_books
+  def display_all_books
+    @books.each_with_index do |book, index|
+      puts "#{index}) Title: '#{book.title}', Author: '#{book.author}'"
+    end
+  end
+end
